修改配置区域

+++ hpool2nossd.py
@@ -1,7 +1,17 @@
+#!/usr/bin/env python3
+# coding=utf-8
+
+'''
+@File  : hpool2nossd.py
+@author: liuyuansharp
+@Time  : 2023/01/10 21:23
+'''
 import os
+import sys
 import subprocess
 import shutil
 import stat
+from time import sleep
 
 from pathlib import Path
 from typing import Dict
@@ -31,31 +41,56 @@ class DriveInfo():
         self.finalizing_flag = False
 
 
-class Hpool2Nossd():
+class hpool2nossd():
     def __init__(self) -> None:
 
-        ###################配置区域###################
+        ###################配置区域开始###################
+        #是否fpt文件优先
         self.fpt_priority = True
-        self.delete_plot_per_time = 1
-        self.spt_space = 88.1  # gb,C5
-        self.fpt_space = 78.1  # gb,C5
+        #每次删除hpool图数量
+        self.delete_plots_num_per_time = 1
         
+        #磁盘挂载根目录
         self.drive_root_path = Path("/srv/")
+        #磁盘文件夹标识（用于排除磁盘挂载根目录下排除无效目录，可为填""，则不作排除）
         self.drive_character = "disk"
-
-        self.plots_dir = "chiapp-files"
-        self.nossd_dir = "nossd"
         
+        #hpool图目录名
+        self.plots_dir = "chiapp-files"
+        #nossd图目录名
+        self.nossd_dir = "nossd"
+        #hpool服务名
+        self.hpool_service = "hpoolpp"
+        #nossd服务名
+        self.nossd_service = "nossd"
+        
+        #nossd安装文件目录
         self.nossd_path = Path("/root/install/nossd-1.2/")
-        self.nossd_start_sh_name = "start_dev.sh"
-        self.nossd_address = "xch1m49h6ny95xgs5a3p2wg6ghnr3vejsqq2pwklq9ae8kg8wgkfujcs26djuq"
+        #nossd脚本名
+        self.nossd_start_sh_name = "start.sh"
+        
+        #nossd压缩等级
         self.nossd_type = 5
+        #nossd机器名
         self.nossd_name = "bsh_001"
-        self.nossd_tmp_drive_paths = ["/srv/dev-disk-by-uuid-9802c526-c5e2-44a4-9e29-b7e1b7b805a0",
-                                    "/srv/dev-disk-by-uuid-ffeb0e19-8f2a-453a-abec-9aa7884c1124",
-                                    "/srv/dev-disk-by-uuid-0ee42af9-6cc1-41a3-992c-c7a80a764b01"]
-        ###################配置区域###################
-    
+        #nossd收益地址
+        self.nossd_address = "xch1m49h6ny95xgs5a3p2wg6ghnr3vejsqq2pwklq9ae8kg8wgkfujcs26djuq"
+        #nossd临时文件磁盘
+        self.nossd_tmp_drive_paths = \
+            [
+                "/srv/dev-disk-by-uuid-9802c526-c5e2-44a4-9e29-b7e1b7b805a0",
+                "/srv/dev-disk-by-uuid-ffeb0e19-8f2a-453a-abec-9aa7884c1124",
+                "/srv/dev-disk-by-uuid-0ee42af9-6cc1-41a3-992c-c7a80a764b01"
+            ]
+            
+        #等待检查间隔(s)
+        self.waitting_time = 300 
+        ###################配置区域结束####################
+        
+        
+        self.spt_space = [96.1,94.1,92.1,90.1,88.1][self.nossd_type - 1]  # gb
+        self.fpt_space = [84.5,82.9,81.3,79.7,78.1][self.nossd_type - 1]  # gb
+        
         self.total_tmp_space = 225
         self.nossd_client = self.nossd_path / "client"
         self.nossd_start_sh = self.nossd_path / self.nossd_start_sh_name
@@ -72,30 +107,76 @@ class Hpool2Nossd():
 
         self.debug = True
 
-    @staticmethod
-    def delete_plots(hard, n):
-        count = 1
-        file = os.listdir(hard)
-
-        print("delete " + hard + "\n")
-        for f in file:
-            plot = os.path.join(hard, f)
-
-            print("deleting " + f + "...")
-
-            if os.path.exists(plot):
-                os.remove(plot)
-                count += 1
-
-            print("done " + f + "...\n")
-
-            if count > n:
-                print("done, deleted " + str(n) + " plots!\n")
+    def reduce_plots(self):
+        deleted_num = 0
+        need_to_delete_num = self.delete_plots_num_per_time
+        
+        for d in self.tmp_spt_or_fpt_drives:
+            drive_info = self.tmp_spt_or_fpt_drives[d]
+            
+            n = 1
+            if drive_info.plots_n >= need_to_delete_num:
+                n = need_to_delete_num
+            else:
+                n = drive_info.plots_n
+                
+            self.delete_plots(drive_info.plots_path, n)
+            deleted_num += n
+            need_to_delete_num -= n
+            drive_info.plots_n -= n
+            self.print_drive_info("deleted", drive_info)
+            
+            if need_to_delete_num == 0:
                 return
-
+                
+        for d in self.spt_or_fpt_drives:
+            drive_info = self.spt_or_fpt_drives[d]
+            
+            n = 1
+            if drive_info.plots_n >= need_to_delete_num:
+                n = need_to_delete_num
+            else:
+                n = drive_info.plots_n
+                
+            self.delete_plots(drive_info.plots_path, n)
+            deleted_num += n
+            need_to_delete_num -= n
+            drive_info.plots_n -= n
+            self.print_drive_info("deleted", drive_info)
+            
+            if need_to_delete_num == 0:
+                return               
+            
     @staticmethod
-    def stop_hpool_service():
-        p = subprocess.Popen("service hpoolpp stop",
+    def delete_plots(plots_path: Path, n: int):
+        count = 1
+        if plots_path.exists():
+            plots = os.listdir(plots_path)
+            for p in plots:
+                plot = os.path.join(plots_path, p)
+                
+                print("deleting " + plot + "...")
+                if os.path.exists(plot):
+                    os.remove(plot)
+                    count += 1
+                print("done " + plot + "...\n")
+
+                if count > n:
+                    print("done, deleted {} plots in {}!\n".format(n,plots_path))
+                    return
+                
+    def is_all_drives_plots_empty(self):
+        count = 0
+        for d in self.all_dirves:
+            count += self.all_dirves[d].plots_n
+            
+        if count == 0:
+            return True
+        
+        return False
+    
+    def set_hpool_service(self,cmd):
+        p = subprocess.Popen("service {} {}}".format(self.hpool_service,cmd),
                              shell=True,
                              stdout=subprocess.PIPE,
                              stderr=subprocess.STDOUT,
@@ -108,9 +189,8 @@ class Hpool2Nossd():
 
         return False
 
-    @staticmethod
-    def start_hpool_service():
-        p = subprocess.Popen("service hpoolpp start",
+    def set_nossd_service(self,cmd):
+        p = subprocess.Popen("service {} {}}".format(self.nossd_service,cmd),
                              shell=True,
                              stdout=subprocess.PIPE,
                              stderr=subprocess.STDOUT,
@@ -120,74 +200,6 @@ class Hpool2Nossd():
 
         if p.returncode == 0:
             return True
-
-        return False
-
-    @staticmethod
-    def stop_nossd_service():
-        p = subprocess.Popen("service nossd stop",
-                             shell=True,
-                             stdout=subprocess.PIPE,
-                             stderr=subprocess.STDOUT,
-                             encoding='utf-8'
-                             )
-        p.wait()
-
-        if p.returncode == 0:
-            return True
-
-        return False
-
-    @staticmethod
-    def start_nossd_service():
-        p = subprocess.Popen("service nossd start",
-                             shell=True,
-                             stdout=subprocess.PIPE,
-                             stderr=subprocess.STDOUT,
-                             encoding='utf-8'
-                             )
-        p.wait()
-
-        if p.returncode == 0:
-            return True
-
-        return False
-
-    def get_nossd_status(self):
-        p = subprocess.Popen("service nossd status",
-                             shell=True,
-                             stdout=subprocess.PIPE,
-                             stderr=subprocess.STDOUT,
-                             encoding='utf-8'
-                             )
-        p.wait()
-
-        if p.returncode == 0:
-            self.nossd_status = p.stdout.readlines()
-
-        if self.debug:
-            for line in self.nossd_status:
-                print(line)
-
-    def is_nossd_farming(self):
-        if not self.is_nossd_plotting() and not self.is_nossd_finalizing():
-            for line in self.nossd_status:
-                if "running" in line:
-                    return True
-
-        return False
-
-    def is_nossd_plotting(self):
-        for line in self.nossd_status:
-            if "Plotting" in line:
-                return True
-
-        return False
-
-    def is_nossd_finalizing(self):
-        for line in self.nossd_status:
-            if "Finalizing" in line:
-                return True
 
         return False
 
@@ -249,7 +261,10 @@ class Hpool2Nossd():
             drive_dir_path = self.drive_root_path / drive_dir
             drive_info = self.get_drive_info(drive_dir_path)
 
-            if self.drive_character in drive_dir:
+            if self.drive_character:
+                if self.drive_character in drive_dir:
+                    self.all_dirves[drive_dir_path] = drive_info
+            else:
                 self.all_dirves[drive_dir_path] = drive_info
 
         return self.all_dirves
@@ -265,7 +280,7 @@ class Hpool2Nossd():
         return no
 
     def print_drive_info(self, status, d: DriveInfo) -> None:
-        print("status: {}".format(status))
+        print("\nstatus: {}".format(status))
         print("drive: {}".format(d.drive_path))
         print(
             "drive info: [total_gb/used_gb/free_gb] : [{}/{}/{}]".format(d.total_gb, d.used_gb, d.free_gb))
@@ -274,24 +289,17 @@ class Hpool2Nossd():
         print("nossd info: fpts_n: {} spts_n: {}".format(d.fpts_n, d.spts_n))
 
         print("plots dir: {}".format(d.plots_path))
-        print("plots info: plots_n: {}".format(d.plots_n))
+        print("plots info: plots_n: {}\n".format(d.plots_n))
+        sys.stdout.flush()
+
 
     def is_completed_drive(self, d: DriveInfo) -> bool:
 
         condition = d.fpts_n >= d.target_fpts_n if self.fpt_priority else d.spts_n >= d.target_spts_n
 
         if not d.plotting_flag and not d.finalizing_flag and condition:
-            self.print_drive_info("completed", d)
+            # self.print_drive_info("completed", d)
             return True
-
-        return False
-
-    def is_ready_drive(self, d: DriveInfo) -> bool:
-
-        if not d.plotting_flag and not d.finalizing_flag:
-            if d.spts_n == 0 and d.fpts_n == 0:
-                self.print_drive_info("ready", d)
-                return True
 
         return False
 
@@ -314,15 +322,7 @@ class Hpool2Nossd():
     def is_tmp_drive(self, d: DriveInfo) -> bool:
 
         if d.tmp_drive_flag:
-            self.print_drive_info("tmp_drive", d)
-            return True
-
-        return False
-    
-    def is_standby_drive(self, d: DriveInfo) -> bool:
-
-        if d.tmp_drive_flag:
-            self.print_drive_info("tmp_drive", d)
+            # self.print_drive_info("tmp_drive", d)
             return True
 
         return False
@@ -394,16 +394,45 @@ class Hpool2Nossd():
             os.chmod(self.nossd_start_sh, stat.S_IRWXU)
 
     def run(self):
-        pass
+        
+        #获取初始磁盘列表
+        self.get_all_dirves()
+        #更新磁盘状态
+        self.get_drives_status()
+        
+        if len(self.plotting_drives) == 0 and len(self.finalizing_drives) == 0:
+            #重启nossd，更新nossd脚本
+            if self.stop_nossd_service():
+                self.update_nossd_start_sh()
+                self.start_nossd_service()
+                        
+        while True:
+            print("waitting {} s ,check drives status again....\n".format(self.waitting_time))
+            sys.stdout.flush()
+            sleep(self.waitting_time) #5 min
+            
+            #更新磁盘状态
+            print("check drives status....\n")
+            sys.stdout.flush()
+            self.get_drives_status()
 
-
+            if len(self.plotting_drives) == 0 and len(self.finalizing_drives) == 0:
+                if not self.is_all_drives_plots_empty():
+                    #重启hpool，删除plots
+                    if self.set_hpool_service("stop"):
+                        self.reduce_plots()
+                        self.set_hpool_service("start")
+                    
+                    #重启nossd，更新nossd脚本
+                    if self.set_nossd_service("stop"):
+                        self.update_nossd_start_sh()
+                        self.set_nossd_service("start")
+                else:
+                    print("done")
+                    return
+                        
 if __name__ == '__main__':
 
-    n = Hpool2Nossd()
-
-    n.get_all_dirves()
-    n.get_drives_status()
+    n = hpool2nossd()
+    n.run()
     
-    n.update_nossd_start_sh()
-    
-    print()
